+++ lib/sec_id/fisn.rb
@@ -36,15 +36,12 @@ module SecId
     # @return [String, nil] the abbreviated instrument description (after the slash)
     attr_reader :description
 
-    has_check_digit false
-
     # @param fisn [String] the FISN string to parse
     def initialize(fisn)
       fisn_parts = parse(fisn)
       @identifier = fisn_parts[:identifier]
       @issuer = fisn_parts[:issuer]
       @description = fisn_parts[:description]
-      @check_digit = nil
     end
 
     # @return [String]